Reorganization of Status Effects (#1187)
From my understanding, a regular status effect is directly mentioned in
the description as a thing.
- Frozen was mentioned by Powder Snow Bucket as its own thing, thus, a
regular one.
- Potion of Slowness' description only says to disable the secondary
effect. Slowness effect should be a system status effect.
- Same for weakness, but this time, as a mechanic, it's not even
inflicting a hermit or a player with weakness.

And quite a few were just listed in the wrong section of the index file.